+++ babel.config.js
@@ -1,4 +1,5 @@
 const path = require('path');
+const isProd = process.env.NODE_ENV === 'production';
 
 module.exports = {
 	presets: [
@@ -15,8 +16,8 @@ module.exports = {
 		[
 			'minify',
 			{
-				removeConsole: true,
-				removeDebugger: true,
+				removeConsole: isProd,
+				removeDebugger: isProd,
 				keepFnName: true,
 				builtIns: false
 			}