+++ site/src/pages/dev/index.vue
@@ -4,6 +4,7 @@
 			<layout-test />
 			<card-test />
 			<icon-test />
+			<avatar-test />
 		</lp-container>
 	</lp-layout>
 </template>
@@ -12,12 +13,14 @@
 import layout from './layout/layout-test.vue';
 import card from './card/card-test.vue';
 import icon from './icon/icon-test.vue';
+import avatar from './avatar/avatar-test.vue';
 
 export default {
 	components: {
 		'layout-test': layout,
 		'card-test': card,
-		'icon-test': icon
+		'icon-test': icon,
+		'avatar-test': avatar
 	}
 };
 </script>