+++ src/components/avatar/index.js
@@ -1,0 +1,7 @@
+import LpAvatar from './src/avatar';
+
+LpAvatar.install = Vue => {
+	Vue.component(LpAvatar.name, LpAvatar);
+};
+
+export default LpAvatar;